+++ src/main/kotlin/me/aguo/plugin/oldyoungradio/ui/ToolWindow.kt
@@ -1,32 +1,35 @@
 package me.aguo.plugin.oldyoungradio.ui
 
-import com.intellij.openapi.actionSystem.ActionManager
-import com.intellij.openapi.actionSystem.ActionPlaces
-import com.intellij.openapi.actionSystem.ActionToolbar
-import com.intellij.openapi.actionSystem.DefaultActionGroup
+import com.intellij.ide.DataManager
+import com.intellij.openapi.actionSystem.*
+import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.ui.DialogPanel
 import com.intellij.openapi.ui.SimpleToolWindowPanel
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.ui.CollectionListModel
-import com.intellij.ui.ToolbarDecorator
 import com.intellij.ui.components.JBList
 import com.intellij.ui.components.JBScrollPane
+import com.intellij.ui.content.ContentManagerEvent
+import com.intellij.ui.content.ContentManagerListener
 import com.intellij.ui.layout.panel
-import me.aguo.plugin.oldyoungradio.TOOL_WINDOW_ROOMS
-import me.aguo.plugin.oldyoungradio.action.AddRoom
-import me.aguo.plugin.oldyoungradio.action.RefreshAllRoom
-import me.aguo.plugin.oldyoungradio.action.RemoveRoom
-import me.aguo.plugin.oldyoungradio.action.StopRadio
-import me.aguo.plugin.oldyoungradio.getAllIds
+import me.aguo.plugin.oldyoungradio.*
+import me.aguo.plugin.oldyoungradio.action.*
 import me.aguo.plugin.oldyoungradio.model.RoomModel
 import me.aguo.plugin.oldyoungradio.network.BiliBiliApi
 import me.aguo.plugin.oldyoungradio.service.StatusService
+import javax.swing.JPanel
 
 
 @Suppress("UnstableApiUsage")
-class ToolWindow : ToolWindowFactory {
+class ToolWindow : ToolWindowFactory, DumbAware {
+    companion object {
+        private const val TAB_SUBSCRIBED_ID = 1
+        private const val TAB_CHANNELS_ID = 2
+        private const val TAB_SUBSCRIBED_NAME = "已订阅"
+        private const val TAB_CHANNELS_NAME = "热门"
+    }
+
     @Suppress("UNCHECKED_CAST")
     private val rooms = BiliBiliApi.getStatusInfoByUids(getAllIds("uid") as List<Int>)
 
@@ -35,48 +38,87 @@ class ToolWindow : ToolWindowFactory {
         StatusService.instance.start()
     }
 
+    class ChannelTabSelectedListener : ContentManagerListener {
+        override fun selectionChanged(event: ContentManagerEvent) {
+            if (CURRENT_CHANNEL != Channels.CHANNEL_ALL && event.content.isSelected && event.content.tabName == TAB_CHANNELS_NAME) {
+                val actionManager = ActionManager.getInstance()
+                val action = actionManager.getAction(CURRENT_CHANNEL.name)
+                action.actionPerformed(
+                    AnActionEvent(
+                        null,
+                        DataManager.getInstance().getDataContext(event.content.component),
+                        ActionPlaces.TOOLWINDOW_POPUP,
+                        Presentation(), actionManager, 0
+                    )
+                )
+            }
+            super.selectionChanged(event)
+        }
+    }
+
     override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
         val contentManager = toolWindow.contentManager
 
-        val simplePanel = SimpleToolWindowPanel(true, true)
-        simplePanel.toolbar = createActionToolbar(simplePanel).component
-        simplePanel.setContent(radios(TOOL_WINDOW_ROOMS))
+        val subscribePanel = SimpleToolWindowPanel(true, true).apply {
+            toolbar = createRoomActionToolbar(this).component
+            setContent(radios(TOOL_WINDOW_ROOMS, TAB_SUBSCRIBED_ID))
+        }
 
-        //TODO: toolbar https://plugins.jetbrains.com/docs/intellij/lists-and-trees.html#toolbardecorator
-        val toolbar = ToolbarDecorator.createDecorator(JBList(rooms))
-        toolbar.setAddAction(AddRoom())
-        // TODO
+        val channelsPanel = SimpleToolWindowPanel(true, true).apply {
+            toolbar = createHotChannelsToolbar(this).component
+            setContent(radios(CHANNEL_TAB_ROOMS, TAB_CHANNELS_ID))
+        }
 
-        val content = contentManager.factory.createContent(simplePanel, null, false)
+//        val label = JLabel("Loading...", AnimatedIcon.Default(), SwingConstants.LEFT)
+//        val panel = panel {
+//        }
+//        panel.add(label)
+//        simplePanel2.setContent(panel)
+
+
+        val content = contentManager.factory.createContent(subscribePanel, TAB_SUBSCRIBED_NAME, false)
+        val content2 = contentManager.factory.createContent(channelsPanel, TAB_CHANNELS_NAME, false)
         contentManager.addContent(content)
+        contentManager.addContent(content2)
+        contentManager.addContentManagerListener(ChannelTabSelectedListener())
     }
-
 }
 
-@Suppress("UnstableApiUsage")
-fun radios(rooms: CollectionListModel<RoomModel>): DialogPanel {
+fun radios(rooms: CollectionListModel<RoomModel>, tabId: Int): JPanel {
     val jbList = JBList(rooms)
-    jbList.addMouseListener(CustomMouseAdapter())
-    jbList.cellRenderer = CustomRenderer()
+    jbList.addMouseListener(CustomMouseAdapter(tabId))
+    jbList.cellRenderer = CustomRenderer(tabId)
     val roomsComponent = JBScrollPane(jbList)
-
     return panel {
-        titledRow("已订阅") {
-            row {
-                roomsComponent()
-            }
+        row {
+            roomsComponent()
         }
     }
 }
 
 
-fun createActionToolbar(panel: SimpleToolWindowPanel): ActionToolbar {
+fun createRoomActionToolbar(panel: SimpleToolWindowPanel): ActionToolbar {
     val actionGroup = DefaultActionGroup()
-    actionGroup.add(RefreshAllRoom())
-    actionGroup.add(AddRoom())
-    actionGroup.add(RemoveRoom())
-    actionGroup.addSeparator()
-    actionGroup.add(StopRadio())
+    actionGroup.apply {
+        add(RefreshAllRoom())
+        add(AddRoom())
+        add(RemoveRoom())
+        addSeparator()
+        add(StopRadio())
+    }
+
+    val actionBar = ActionManager.getInstance().createActionToolbar(ActionPlaces.TOOLBAR, actionGroup, true)
+    actionBar.setTargetComponent(panel)
+    return actionBar
+}
+
+fun createHotChannelsToolbar(panel: SimpleToolWindowPanel): ActionToolbar {
+    val actionGroup = DefaultActionGroup()
+    actionGroup.apply {
+        add(GenerateChannelsComboBox())
+        addSeparator()
+        add(StopRadio())
+    }
     val actionBar = ActionManager.getInstance().createActionToolbar(ActionPlaces.TOOLBAR, actionGroup, true)
     actionBar.setTargetComponent(panel)
     return actionBar